+++ examples/c00-readme.rs
@@ -60,7 +60,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
 		println!("\n--- Answer:");
 		let chat_res = client.exec_chat(model, chat_req.clone(), None).await?;
-		println!("{}", chat_res.content.as_deref().unwrap_or("NO ANSWER"));
+		println!("{}", chat_res.content_text_as_str().unwrap_or("NO ANSWER"));
 
 		println!("\n--- Answer: (streaming)");
 		let chat_res = client.exec_chat_stream(model, chat_req.clone(), None).await?;