+++ src/chat/mod.rs
@@ -7,6 +7,7 @@ mod chat_options;
 mod chat_req;
 mod chat_res;
 mod chat_stream;
+mod message_content;
 mod tool;
 
 // -- Flatten
@@ -14,6 +15,7 @@ pub use chat_options::*;
 pub use chat_req::*;
 pub use chat_res::*;
 pub use chat_stream::*;
+pub use message_content::*;
 pub use tool::*;
 
 // endregion: --- Modules